requirements.txt
streamlit
torch
torchvision
pandas
numpy
scikit-learn
Pillow
